fix(runtime): preserve egress on correction turns
